+++ pyhdx/__init__.py
@@ -4,9 +4,8 @@ from pyhdx.models import (
     Coverage,
     HDXMeasurementSet,
 )
-from pyhdx.fileIO import read_dynamx
+from pyhdx.datasets import read_dynamx
 from pyhdx.fitting_torch import TorchFitResult, TorchFitResultSet
-from pyhdx.batch_processing import StateParser
 from pyhdx.__version__ import __version__
 
 VERSION_STRING = f"PyHDX {__version__}"
